Revert "Update api.class.php"
This reverts commit 95074290b173d764cc1d8812f29900dc3c27b788.

+++ lib/class/api.class.php
@@ -717,7 +717,7 @@ class Api
         if (!self::check_parameter($input, array('filter'), 'artist')) {
             return false;
         }
-        $uid     = (int) scrub_in($input['filter']);
+        $uid     = scrub_in($input['filter']);
         $user    = User::get_from_username(Session::username($input['auth']));
         $include = (is_array($input['include'])) ? $input['include'] : explode(',', $input['include']);
         switch ($input['format']) {
